+++ tests/integration/cases/bool_list/CSharp_type_hints_safe_seq_array@v10.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 var my_data = new bool[] {
     true,
     false,