+++ tests/integration/cases/dict_with_list_value/CSharp_heterogeneous_strategy_record_list_val@v10.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System;
 record Record0(string Name, int[] Scores);
 class Check {
     public static void Main() {