+++ tests/integration/cases/mixed_number_list/CSharp_modifiers_readonly_mixed_numbers@v10.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 class Check {
 readonly double[] my_data = new double[] {
     1,