+++ tests/integration/cases/nested_sequences/CSharp_heterogeneous_strategy_record@v10.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 class Check {
     public static void Main() {
 var my_data = new int[][][] {