+++ tests/integration/cases/simple_sequence/CSharp_sequence_array_varname@v10.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 var my_data = new object[] {
     1,
     "hello",